+++ app/src/main/java/open/furaffinity/client/adapter/msgOthersSectionsPagerAdapter.java
@@ -10,7 +10,7 @@ import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentPagerAdapter;
 
 import open.furaffinity.client.R;
-import open.furaffinity.client.fragmentTabsOld.msgOthersList;
+import open.furaffinity.client.fragmentTabsNew.msgOthersList;
 import open.furaffinity.client.utilities.messageIds;
 
 public class msgOthersSectionsPagerAdapter extends FragmentPagerAdapter {
