--- app/src/main/java/open/furaffinity/client/fragmentTabsOld/userJournals.java
@@ -1,116 +1,0 @@
-package open.furaffinity.client.fragmentTabsOld;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.Toast;
-
-import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import open.furaffinity.client.R;
-import open.furaffinity.client.abstractClasses.abstractPage;
-import open.furaffinity.client.adapter.journalListAdapter;
-import open.furaffinity.client.listener.EndlessRecyclerViewScrollListener;
-import open.furaffinity.client.pages.journals;
-import open.furaffinity.client.utilities.messageIds;
-
-public class userJournals extends Fragment {
-    private LinearLayoutManager layoutManager;
-
-    private RecyclerView recyclerView;
-    private RecyclerView.Adapter mAdapter;
-    private EndlessRecyclerViewScrollListener endlessRecyclerViewScrollListener;
-
-    private journals page;
-
-    private int loadingStopCounter = 3;
-    private boolean isLoading = false;
-    private List<HashMap<String, String>> mDataSet = new ArrayList<>();
-
-    private void getElements(View rootView) {
-        layoutManager = new LinearLayoutManager(getActivity());
-
-        recyclerView = rootView.findViewById(R.id.recyclerView);
-    }
-
-    private void fetchPageData() {
-        if(!isLoading && loadingStopCounter > 0) {
-            isLoading = true;
-            page = new journals(page);
-            page.execute();
-        }
-    }
-
-    private void initPages() {
-        recyclerView.setLayoutManager(layoutManager);
-        mAdapter = new journalListAdapter(mDataSet, getActivity());
-        recyclerView.setAdapter(mAdapter);
-
-        page = new journals(getActivity(), new abstractPage.pageListener() {
-            @Override
-            public void requestSucceeded(abstractPage abstractPage) {
-                List<HashMap<String, String>> pageResults = page.getPageResults();
-
-                int curSize = mAdapter.getItemCount();
-
-                if (pageResults.size() == 0 && loadingStopCounter > 0) {
-                    loadingStopCounter--;
-                }
-
-                //Deduplicate results
-                List<String> newPostPaths = pageResults.stream().map(currentMap -> currentMap.get("journalPath")).collect(Collectors.toList());
-                List<String> oldPostPaths = mDataSet.stream().map(currentMap -> currentMap.get("journalPath")).collect(Collectors.toList());
-                newPostPaths.removeAll(oldPostPaths);
-                pageResults = pageResults.stream().filter(currentMap -> newPostPaths.contains(currentMap.get("journalPath"))).collect(Collectors.toList());
-                mDataSet.addAll(pageResults);
-                mAdapter.notifyItemRangeInserted(curSize, mDataSet.size());
-
-                isLoading = false;
-            }
-
-            @Override
-            public void requestFailed(abstractPage abstractPage) {
-                loadingStopCounter--;
-                isLoading = false;
-                Toast.makeText(getActivity(), "Failed to load data for journals", Toast.LENGTH_SHORT).show();
-            }
-        }, getArguments().getString(messageIds.pagePath_MESSAGE));
-    }
-
-    private void updateUIElementListeners(View rootView) {
-        endlessRecyclerViewScrollListener = new EndlessRecyclerViewScrollListener(layoutManager) {
-            @Override
-            public void onLoadMore(int pageNumber, int totalItemsCount, RecyclerView view) {
-                page.setPage(Integer.toString(page.getPage() + 1));
-                fetchPageData();
-            }
-        };
-
-        //noinspection deprecation
-        recyclerView.setOnScrollListener(endlessRecyclerViewScrollListener);
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View rootView = inflater.inflate(R.layout.fragment_recycler_view, container, false);
-        getElements(rootView);
-        initPages();
-        fetchPageData();
-        updateUIElementListeners(rootView);
-        return rootView;
-    }
-}
